project_evaluator/requirements.txt
# Core Dependencies
GitPython==3.1.41
radon==6.0.1
rich==13.7.0

# Utilities
PyYAML==6.0.1
pathspec==0.12.1

# Code Analysis
lizard==1.17.10
pygments==2.17.2

# Type Checking (Dev)
mypy==1.8.0
types-PyYAML==6.0.12.12

# Testing (Dev)
pytest==7.4.4
pytest-cov==4.1.0
